doc/ms.ms.in: Reform terminology.
"marker" -> "mark"

This is consistent with groff_mm(7), and with refer(1)'s description of
an inline bibliographic citation.  I suppose "mark" has a more flexible
meaning in typography than I had thought.

Also, "mark" is shorter and "-er" buys nothing semantically except
clarity that it's not being used as a verb, which is not a hazard in
this document's specimens, where it's preceded by an article,
quantifier, or other adjective.