[doc,man]: Lean less hard on "transparent" crutch.
"Special" and "transparent" are the most abused words in the troff
documentary corpus.

Special *how*?  Transparent *to what*?

Relatedly (to my jaundiced eye):

“My latest pet writing peeve: the use of "a kind of" and "a sort of."  I
don't know when I began noticing this writing tic, in myself and in the
larger culture, and I don't know how long it has been around.  I'm
certain I've seen these phrases in writing from as long as a century
ago, but it's all over the place now.  I associate it with a sort
of—oops, I did it again!—literary affectation that one sees not only on
social media but in legacy media as well.  Along the lines of Lionel
Trilling saying, "Goethe says somewhere...," but in a tweet, a post, or
a piece.  It evokes a Jamesian spirit of the ineffable, the not quite,
the almost, that liminal space between the is and the is not.  It's
meant to suggest a fineness of distinction, I suppose, the subtlety of
one's antenna.  But it seems like little more than a cover for an
insufficiency of one's distinctions, a failure of thought, an inability
to state as fully as possible what one is thinking or trying to think.
I can't stand it. I'm going to try to stop. Kinda, sorta.”
  -- Corey Robin