[doc,man]: Adjust pedagogy in "Copy mode".
Take a leaf from bash(1)'s playbook and characterize the escape
character as sometimes operating as a _quotation_ character.

In a recursive interpretation process, "escaping" a token refers to
"getting out" of the current interpretation context so that the token is
interpreted in an enclosing one; "quoting" refers to "pushing down" or
"hiding" some input from the current interpretation context so that a
token gets interpreted only in an inferior context.

It's kind of a shame that we use the same input token to represent both
of these processes, which are not only distinguishable but opposite in
meaning.  But, hey, man--it was the seventies.

"I tied an onion to my belt, which was the style at the time."