[doc,man]: Clarify a point.
(So to speak.)

GNU troff's support for sub-integer type size increments doesn't
categorically cause an incompatibility.  It's troublesome only in the
scenario, hard to motivate and unobserved by me in the wild, of legacy
AT&T documents explicitly putting a "u" scaling unit on an argument to
the `ps` request.  (Because, as explained here, AT&T troff ignored
scaling units in `ps`'s arguments, there was no benefit to doing so.)