Revert "[troff]: Add assertion."
This reverts commit 75321a3ab4a779cc542c91cbc304b372774243f4.

I was wrong; the scenario described (hyphenation is attempted on a line
with no nodes in it) _can_ happen.  I happened upon it by accident when
trying to format the "meref.me" document on a terminal while forgetting
to load the me(7) macro package.  This produced a badly corrupted
document; the aforementioned scenario appears to involve use of a
diversion with an indentation greater than the line length.  This likely
is not a practical text formatting scenario, but impractical scenarios
should not crash the formatter.  Delete the assertion.

Fixes:

$ ~/groff-HEAD/bin/groff -T utf8 -a ./build/doc/meref.me
troff: ../src/roff/troff/env.cpp:2171: void environment::possibly_hyphenate_line(bool): Assertion `*startp != 0' failed.
/.../groff-HEAD/bin/groff: error: troff: Aborted (core dumped)