[gdiffmk]: Revise provenance and license notices.
Elsewhere we are at pains to distinguish the status of materials in
"contrib/" as "not being part of groff", so it's confusing to contradict
that inside gdiffmk's own files.

Revise and parallelize these disclosures of provenance.

Stylistically tweak comment blocks involving legal notices to make them
cohere and to distinguish commentary on other subjects.