contrib/sboxes/sboxes.am: Revise legalese.
Update copyright and authorial notices.

Deri committed this file, which is obviously adapted from Bertrand's
Automake scripts for groff, and just as obviously features his own work
in the adaptation.  Authorship's relatively easy to determine.

Determining the precise contents of the copyright notice per the GNU
Maintainers' Guide is less so.

Add copyright notice for myself.  I have never executed copyright
assignment paperwork with the FSF, so my "legally significant"[1]
changes to this file can't be under the FSF's copyright.

Justification (commit ID, summary, net lines added by commit):

2021 67b2e3c8ac [sboxes]: Leap through flaming DRY hoop. (+15)

(I ignored another commit of mine that exceeded the 15-line limit
because it wasn't my work in a copyright sense; I pasted in a license
notice.)

But Ingo contributed to this file, too.  Should he get a copyright
notice?  Answering that question blunders straight into the
hair-splitting questions I've been repeatedly raising in recent commit
messages, which no doubt some readers (if, in fact, I have any) have
scoffed at as implausible.

Here are his changes, slightly summarized.

commit 6bdf48eb6755b429aa82c9c5c532bd4339c1deb1
Author: Ingo Schwarze <schwarze@openbsd.org>
Date:   Sun Apr 3 00:39:34 2022 +0200

    msboxes.ms is documentation, not merely an example
...
 contrib/sboxes/sboxes.am | 30 ++++++++++++++----------------
 1 file changed, 14 insertions(+), 16 deletions(-)

commit 37cacdc36cc9cdcc1025144236f9734621aa99ae
Author: Ingo Schwarze <schwarze@openbsd.org>
Date:   Sat Apr 2 12:56:55 2022 +0200

    Stop installing some *.in files.
...
 contrib/sboxes/sboxes.am | 7 +------
 1 file changed, 1 insertion(+), 6 deletions(-)

commit 32f057cab219de1f44e91aa38436e81ac77a6ae6
Author: Ingo Schwarze <schwarze@openbsd.org>
Date:   Tue Mar 22 15:08:50 2022 +0100

    [sboxes] fix non-portable syntax in the msboxes.pdf target
...
 contrib/sboxes/sboxes.am | 2 +-
 1 file changed, 1 insertion(+), 1 deletion(-)

commit 4e0905d82276cd124ceb6bbb4a582910393aef22
Author: Ingo Schwarze <schwarze@openbsd.org>
Date:   Tue Mar 22 13:41:31 2022 +0100

    [sboxes] make the build of msboxes.{ms,pdf} work with BSD make
...
 contrib/sboxes/sboxes.am | 4 ++--
 1 file changed, 2 insertions(+), 2 deletions(-)

Did Ingo make any single commit to this file that contributed at least
15 lines of code, on net?  No.

Has Ingo inserted more than 15 lines, _not_ counting deletions?
Yes; 14+1+1+2=17.

Ingo's deleted 16+6+1+2=25 lines total.

So, under one of the simplistic "cumulativity" analyses I've proffered
in commit logs--and let's be candid and acknowledge that simplistic
analyses are the first and often only ones engineering managers reach
for--Ingo's "contributed nothing" to this file, a determination that I
find unpersuasive, and moreover insulting to volunteer contributors.

At least one fact is easy to decide: if Ingo has copyrightable changes
in this file, we can date them all to 2022.  Count the small blessings.

I'm leaving out a notice for Ingo for the time being because I cannot
justify it on the basis of the guidance in the GNU Maintainers' Guide.

In my opinion, that document should offer more guidance on matters like
this.  But I'm disinclined to spend further cycles on trying to develop
such guidance (and which I'd need a copyright attorney on tap to consult
with anyway) without pecuniary incentive.  Deciding (or punting on)
this sort of issue is not a rewarding aspect of a GNU maintainer's job.

None of the foregoing derogates Ingo's contribution to groff or to this
file, specifically.  I value portable make(1) files and I'm sure some of
our users do, too.  We just need to get pattern rules into BSD makes. ;)

[1] "A change of just a few lines (less than 15 or so) is not legally
significant for copyright."  Conversely, >= 15 lines would be.

https://www.gnu.org/prep/maintain/html_node/Legally-Significant.html