[chem]: Simplify `--version` output.
* contrib/chem/chem.pl Drop `copyright` scalar; see below.  Bump
  revision or "micro" portion of version number.

  (version): Stop reporting copyright and license notice after version
  disclosure.  No other program groff installs, not even others authored
  by Bernd Warken, handles `--version` in this way, except for the
  namesake program groff(1) itself.

nroff(1) and pdfmom(1), which _wrap_ groff(1), do cause emission of
these notices when given a `--version` argument, but only by dint of
their function as wrappers.

I see no reason to similarly change `groff --version` output; _as_ the
namesake of the project/package/distribution, a user might reasonably
seek to discover the provenance of an installed groff via that means.